fastapi/requirements.txt
alembic==1.8.1
fastapi==0.88.0
fastapi_jwt==0.1.8
passlib==1.7.4
pydantic==1.9.0
python-dotenv==0.21.0
SQLAlchemy==1.4.35
uvicorn==0.17.6
gunicorn==20.1.0
sqlmodel==0.0.6
pydantic==1.9.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
psycopg2==2.9.3
email-validator==1.2.0


# anyio==3.5.0
# asgiref==3.5.1
# bcrypt==4.0.1
# certifi==2020.6.20
# cffi==1.15.1
# charset-normalizer==2.0.12
# click==8.1.3
# colorama==0.4.4
# cryptography==38.0.4
# dnspython==2.2.1
# docopt==0.6.2
# ecdsa==0.17.0
# email-validator==1.2.0
# fastapi==0.75.2
# fastapi-jwt==0.1.8
# fastapi-jwt-auth==0.5.0
# greenlet==1.1.2
# gunicorn==20.1.0
# h11==0.13.0
# idna==3.3
# jose==1.0.0
# Mako==1.2.3
# MarkupSafe==2.1.1
# mypy==0.991
# mypy-extensions==0.4.3
# passlib==1.7.4
# pipreqs==0.4.11
# psycopg2==2.9.3
# psycopg2-binary==2.9.5
# pyasn1==0.4.8
# pycparser==2.21
# pydantic==1.9.0
# PyJWT==1.7.1
# python-dotenv==0.20.0
# python-jose==3.3.0
# python-multipart==0.0.5
# requests==2.27.1
# rsa==4.8
# six==1.16.0
# sniffio==1.2.0
# SQLAlchemy==1.4.35
# sqlalchemy-stubs==0.4
# sqlalchemy2-stubs==0.0.2a22
# sqlmodel==0.0.6
# starlette==0.17.1
# tomli==2.0.1
# typing_extensions==4.2.0
# urllib3==1.26.9
# uvicorn==0.17.6
# wincertstore==0.2
# yarg==0.1.9
